content/docker/groq/requirements.txt
groq
python-dotenv
jupyterlab
jupyter-ai==3.0.0b9
sqlparse
duckdb
